NFDIV-1834: Refactor App1 Resubmission event so it can be used by solicitor journey as well as citizen journey
