+++ src/main/java/uk/gov/hmcts/divorce/common/event/Applicant1Resubmit.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.divorce.citizen.event;
+package uk.gov.hmcts.divorce.common.event;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,13 +19,14 @@ import java.util.List;
 
 import static uk.gov.hmcts.divorce.divorcecase.model.State.AwaitingApplicant1Response;
 import static uk.gov.hmcts.divorce.divorcecase.model.State.AwaitingApplicant2Response;
+import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.APPLICANT_1_SOLICITOR;
 import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.CITIZEN;
 import static uk.gov.hmcts.divorce.divorcecase.model.access.Permissions.CREATE_READ_UPDATE;
 import static uk.gov.hmcts.divorce.divorcecase.validation.ValidationUtil.validateApplicant1BasicCase;
 
 @Slf4j
 @Component
-public class CitizenApplicant1Resubmit implements CCDConfig<CaseData, State, UserRole> {
+public class Applicant1Resubmit implements CCDConfig<CaseData, State, UserRole> {
 
     public static final String APPLICANT_1_RESUBMIT = "applicant1-resubmit";
 
@@ -41,9 +42,9 @@ public class CitizenApplicant1Resubmit implements CCDConfig<CaseData, State, Use
         configBuilder
             .event(APPLICANT_1_RESUBMIT)
             .forStates(AwaitingApplicant1Response)
-            .name("Applicant 1 Resubmit")
+            .name("Resubmit Applicant 1 Answers")
             .description("Applicant 1 resubmits for joint application")
-            .grant(CREATE_READ_UPDATE, CITIZEN)
+            .grant(CREATE_READ_UPDATE, APPLICANT_1_SOLICITOR, CITIZEN)
             .retries(120, 120)
             .aboutToSubmitCallback(this::aboutToSubmit);
     }
